🔨 move projected data pattern components closer to where they're used
