+++ packages/@ourworldindata/grapher/src/mapCharts/MapComponents.tsx
@@ -21,7 +21,7 @@ import {
 import { isMapRenderFeature } from "./MapHelpers"
 import { getExternalMarkerEndPosition } from "./MapAnnotations"
 import { Patterns } from "../core/GrapherConstants"
-import { DottedProjectedDataPattern } from "../chart/ChartComponents"
+import { calculateLightnessScore } from "../color/ColorUtils"
 
 export function BackgroundCountry<Feature extends RenderFeature>({
     feature,
@@ -176,7 +176,7 @@ export function NoDataPattern({
     )
 }
 
-export function MapProjectedDataPattern({
+export function ProjectedDataPattern({
     color,
     scale = 1,
     forLegend = false,
@@ -192,6 +192,50 @@ export function MapProjectedDataPattern({
             scale={scale}
             dotOpacity={forLegend ? 0.2 : undefined}
         />
+    )
+}
+
+function DottedProjectedDataPattern({
+    patternId,
+    color,
+    scale = 1,
+    patternSize = 4,
+    dotSize = patternSize / 4,
+    dotOpacity,
+}: {
+    patternId: string
+    color: string
+    scale?: number
+    patternSize?: number
+    dotSize?: number
+    dotOpacity?: number // inferred from color lightness if not provided
+}): React.ReactElement {
+    // Choose the dot opacity based on the lightness of the color:
+    // - If the color is light, make the dots more transparent
+    // - If the color is dark, make the dots more opaque
+    const lightness = calculateLightnessScore(color) ?? 0
+    const opacity = dotOpacity ?? Math.max(1 - lightness, 0.1)
+
+    return (
+        <pattern
+            id={patternId}
+            patternUnits="userSpaceOnUse"
+            width={patternSize}
+            height={patternSize}
+            patternTransform={`rotate(45) scale(${scale})`}
+        >
+            {/* colored background */}
+            <rect width={patternSize} height={patternSize} fill={color} />
+
+            {/* dots */}
+            <circle
+                cx={patternSize / 2}
+                cy={patternSize / 2}
+                r={dotSize}
+                fill="black"
+                fillOpacity={opacity}
+            />
+        </pattern>
     )
 }
